requirements.txt
openai-agents>=0.0.10
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
lxml>=5.2.0
yfinance>=0.2.50
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
duckdb>=1.0.0
httpx>=0.27.0
trafilatura>=1.12.0
feedparser>=6.0.11
python-dotenv>=1.0.0
